refactor: move generic meta APIs under kv
# Summary

Move the generic name-value and name-id-value API helpers out of concrete API implementations and into the KV helper module while preserving their public re-export paths.

# Details

These helpers model reusable KV access patterns rather than a concrete metaservice resource API, so keeping them under the KV module better matches their role and avoids mixing generic infrastructure with resource-specific implementations.

The existing top-level module exports remain available for downstream callers; internal users now import the traits through the crate-level re-export.
