+++ src/meta/api/src/lib.rs
@@ -40,8 +40,6 @@ pub(crate) use api_impl::errors;
 pub use api_impl::garbage_collection_api;
 pub use api_impl::index_api;
 pub use api_impl::lock_api;
-pub use api_impl::name_id_value_api;
-pub use api_impl::name_value_api;
 pub use api_impl::ref_api;
 pub(crate) use api_impl::row_access_policy_api;
 pub(crate) use api_impl::schema_api;
@@ -73,6 +71,8 @@ pub use garbage_collection_api::GarbageCollectionApi;
 pub use index_api::IndexApi;
 pub use kv::crud;
 pub use kv::fetch_util as kv_fetch_util;
+pub use kv::name_id_value_api;
+pub use kv::name_value_api;
 pub use kv::pb_api as kv_pb_api;
 pub use kv::pb_crud_api as kv_pb_crud_api;
 pub use kv_app_error::KVAppResultExt;
